Made graph view selectors independent from resource view selectors
