+++ client/app/scripts/components/nodes-resources.js
@@ -5,6 +5,10 @@ import Logo from './logo';
 import ZoomWrapper from './zoom-wrapper';
 import NodesResourcesLayer from './nodes-resources/node-resources-layer';
 import { layersTopologyIdsSelector } from '../selectors/resource-view/layout';
+import {
+  resourcesZoomLimitsSelector,
+  resourcesZoomStateSelector,
+} from '../selectors/resource-view/zoom';
 
 
 class NodesResources extends React.Component {
@@ -24,7 +28,10 @@ class NodesResources extends React.Component {
       <div className="nodes-resources">
         <svg id="canvas" width="100%" height="100%">
           <Logo transform="translate(24,24) scale(0.25)" />
-          <ZoomWrapper svg="canvas" bounded forwardTransform fixVertical>
+          <ZoomWrapper
+            svg="canvas" bounded forwardTransform fixVertical
+            zoomLimitsSelector={resourcesZoomLimitsSelector}
+            zoomStateSelector={resourcesZoomStateSelector}>
             {transform => this.renderLayers(transform)}
           </ZoomWrapper>
         </svg>
